refactor: 예외 발생 로직 리팩터링
