--- backend/src/main/java/com/darass/darass/auth/oauth/exception/AuthenticationException.java
@@ -1,8 +1,0 @@
-package com.darass.darass.auth.oauth.exception;
-
-public class AuthenticationException extends RuntimeException {
-
-    public AuthenticationException(String message) {
-        super(message);
-    }
-}
