+++ backend/src/main/java/com/darass/darass/exception/httpbasicexception/auth/AuthenticationException.java
@@ -1,0 +1,10 @@
+package com.darass.darass.exception.httpbasicexception.auth;
+
+import com.darass.darass.exception.httpbasicexception.CustomException;
+
+public class AuthenticationException extends CustomException {
+
+    public AuthenticationException(String message, Integer code) {
+        super(message, code);
+    }
+}
