+++ backend/src/test/java/com/darass/darass/DarassApplicationTests.java
@@ -1,13 +1,12 @@
 package com.darass.darass;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class DarassApplicationTests {
 
-    @Test
-    void contextLoads() {
-    }
+//    @Test
+//    void contextLoads() {
+//    }
 
 }
