+++ backend/src/test/java/com/darass/darass/auth/oauth/infrastructure/JwtTokenProviderTest.java
@@ -3,7 +3,7 @@ package com.darass.darass.auth.oauth.infrastructure;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import com.darass.darass.auth.oauth.exception.AuthenticationException;
+import com.darass.darass.exception.httpbasicexception.auth.AuthenticationException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
